revert names of Platform classes because of bincompat
