+++ core/js/src/main/scala/scodec/bits/ByteVectorCompanionCrossPlatform.scala
@@ -30,4 +30,18 @@
 
 package scodec.bits
 
-private[bits] trait ByteVectorCompanionScalaVersion
+import scala.scalajs.js.typedarray.{ArrayBuffer, Int8Array, TypedArrayBuffer}
+
+private[bits] trait ByteVectorCompanionCrossPlatform extends ByteVectorPlatform { self: ByteVector.type =>
+  def view(typedArray: Int8Array): ByteVector = ByteVector.view(TypedArrayBuffer.wrap(typedArray))
+
+  def view(arrayBuffer: ArrayBuffer): ByteVector = ByteVector.view(TypedArrayBuffer.wrap(arrayBuffer))
+
+  def fromTypedArray(typedArray: Int8Array): ByteVector = {
+    val copy = new Int8Array(typedArray.length)
+    copy.set(typedArray)
+    ByteVector.view(copy)
+  }
+
+  def fromArrayBuffer(arrayBuffer: ArrayBuffer): ByteVector = ByteVector.fromTypedArray(new Int8Array(arrayBuffer))
+}
