+++ core/jvm/src/main/scala/scodec/bits/ByteVectorCompanionCrossPlatform.scala
@@ -30,16 +30,5 @@
 
 package scodec.bits
 
-// import scala.util.FromDigits
-
-private[bits] trait ByteVectorCompanionScalaVersion { self: ByteVector.type =>
-  /* Disabled because FromDigits is only available in scalac snapshots/nightlies as of https://github.com/lampepfl/dotty/pull/11852
-  given FromDigits.WithRadix[ByteVector] with {
-    def fromDigits(digits: String, radix: Int): ByteVector =
-      radix match {
-        case 16 => ByteVector.fromValidHex(digits)
-        case _  => ByteVector.fromValidHex(new java.math.BigInteger(digits, radix).toString(16))
-      }
-  }
-  */
+private[bits] trait ByteVectorCompanionCrossPlatform extends ByteVectorPlatform { self: ByteVector.type =>
 }
