+++ core/shared/src/main/scala-2/scodec/bits/BitVectorPlatform.scala
@@ -30,15 +30,4 @@
 
 package scodec.bits
 
-// import scala.util.FromDigits
-
-private[bits] trait BitVectorCompanionScalaVersion { self: BitVector.type =>
-  /* Disabled because FromDigits is only available in scalac snapshots/nightlies as of https://github.com/lampepfl/dotty/pull/11852
-  given FromDigits.WithRadix[BitVector] with {
-    def fromDigits(digits: String, radix: Int): BitVector = radix match {
-      case 16 => ByteVector.fromValidHex(digits).bits
-      case _ => ByteVector.fromValidHex(new java.math.BigInteger(digits, radix).toString(16)).bits
-    }
-  }
-  */
-}
+private[bits] trait BitVectorPlatform
