+++ core/shared/src/main/scala-3/scodec/bits/ByteVectorPlatform.scala
@@ -30,5 +30,16 @@
 
 package scodec.bits
 
-private[bits] trait ByteVectorCompanionPlatform extends ByteVectorCompanionScalaVersion { self: ByteVector.type =>
+// import scala.util.FromDigits
+
+private[bits] trait ByteVectorPlatform { self: ByteVector.type =>
+  /* Disabled because FromDigits is only available in scalac snapshots/nightlies as of https://github.com/lampepfl/dotty/pull/11852
+  given FromDigits.WithRadix[ByteVector] with {
+    def fromDigits(digits: String, radix: Int): ByteVector =
+      radix match {
+        case 16 => ByteVector.fromValidHex(digits)
+        case _  => ByteVector.fromValidHex(new java.math.BigInteger(digits, radix).toString(16))
+      }
+  }
+  */
 }
