+++ core/shared/src/main/scala/scodec/bits/BitVector.scala
@@ -1306,7 +1306,7 @@ sealed abstract class BitVector
   * @groupname base Base Conversions
   * @groupprio base 3
   */
-object BitVector extends BitVectorCompanionPlatform {
+object BitVector extends BitVectorCompanionCrossPlatform {
 
   /** Empty bit vector.
     * @group constants
